Test proposed change!
Test the new URL before applying it to the sources file.

This is to work around mirrors which are temporarily down. The idea is to check if
the distribution exists on the 'next' mirror. E.g. currently it's archive.ubuntu.com,
so we test if the distribution moved to old-releases.ubuntu.com.

If it's an arbitrary Ubuntu mirror, we test archive.ubuntu.com next.

If there's an error with old-releases.ubuntu.com, we should probably assume the
mirror is overloaded or only temporarily down (FML).